+++ test/rails_app/app/models/active_model_object.rb
@@ -1,0 +1,14 @@
+class ActiveModelObject
+  include ActiveModel::Model
+  include SpreadsheetArchitect
+
+  attr_accessor :name, :content, :created_at
+
+  def spreadsheet_columns
+    [
+      :name,
+      :content,
+      :created_at
+    ]
+  end
+end