+++ test/rails_app/app/models/bad_plain_ruby_object.rb
@@ -1,0 +1,3 @@
+class BadPlainRubyObject
+  include SpreadsheetArchitect
+end