+++ test/rails_app/app/models/custom_post.rb
@@ -1,0 +1,15 @@
+class CustomPost < ActiveRecord::Base
+  self.table_name = :posts
+
+  include SpreadsheetArchitect
+
+  def spreadsheet_columns
+    [
+      :name,
+      ['The Content', :content],
+      :created_at, 
+      ['Created At 2', created_at.strftime("%Y-%m-%d")],
+      [:asd, 'tadaaa']
+    ]
+  end
+end