+++ test/rails_app/db/migrate/20170103234524_add_posts.rb
@@ -1,0 +1,10 @@
+class AddPosts < ActiveRecord::Migration[5.0]
+  def change
+    create_table :posts do |t|
+      t.string :name
+      t.text :content
+      t.integer :age
+      t.timestamps
+    end
+  end
+end